+++ src/test/java/address/TestApp.java
@@ -1,11 +1,10 @@
 package address;
 
 import address.model.UserPrefs;
-import address.model.datatypes.AddressBook;
 import address.model.datatypes.ReadOnlyAddressBook;
 import address.storage.StorageAddressBook;
 import address.util.Config;
-import address.util.TestUtil;
+import address.testutils.TestUtil;
 
 import java.util.function.Supplier;
