--- src/test/java/address/unittests/util/XmlUtilTest.java
@@ -1,95 +1,0 @@
-package address.unittests.util;
-
-import address.exceptions.DataConversionException;
-import address.model.datatypes.AddressBook;
-import address.storage.StorageAddressBook;
-import address.util.AddressBookBuilder;
-import address.util.FileUtil;
-import address.util.TestUtil;
-import address.util.XmlUtil;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-
-import static org.junit.Assert.assertEquals;
-
-public class XmlUtilTest {
-
-    private static final String TEST_DATA_FOLDER = FileUtil.getPath("src/test/data/XmlUtilTest/");
-    private static final File EMPTY_FILE = new File(TEST_DATA_FOLDER + "empty.xml");
-    private static final File MISSING_FILE = new File(TEST_DATA_FOLDER + "missing.xml");
-    private static final File VALID_FILE = new File(TEST_DATA_FOLDER + "validAddressBook.xml");
-    private static final File TEMP_FILE = new File(TestUtil.appendToSandboxPath("tempAddressBook.xml"));
-
-    @Rule
-    public ExpectedException thrown = ExpectedException.none();
-
-    @Test
-    public void getDataFromFile_nullFile_AssertionError() throws Exception {
-        thrown.expect(AssertionError.class);
-        XmlUtil.getDataFromFile(null, AddressBook.class);
-    }
-
-    @Test
-    public void getDataFromFile_nullClass_AssertionError() throws Exception {
-        thrown.expect(AssertionError.class);
-        XmlUtil.getDataFromFile(VALID_FILE, null);
-    }
-
-    @Test
-    public void getDataFromFile_missingFile_FileNotFoundException() throws Exception {
-        thrown.expect(FileNotFoundException.class);
-        XmlUtil.getDataFromFile(MISSING_FILE, AddressBook.class);
-    }
-
-    @Test
-    public void getDataFromFile_emptyFile_DataFormatMismatchException() throws Exception {
-        thrown.expect(DataConversionException.class);
-        XmlUtil.getDataFromFile(EMPTY_FILE, AddressBook.class);
-    }
-
-    @Test
-    public void getDataFromFile_validFile_validResult() throws Exception {
-        StorageAddressBook dataFromFile = XmlUtil.getDataFromFile(VALID_FILE, StorageAddressBook.class);
-        assertEquals(2, dataFromFile.getPersonList().size());
-        assertEquals(1, dataFromFile.getTagList().size());
-    }
-
-    @Test
-    public void saveDataToFile_nullFile_AssertionError() throws Exception {
-        thrown.expect(AssertionError.class);
-        XmlUtil.saveDataToFile(null, new AddressBook());
-    }
-
-    @Test
-    public void saveDataToFile_nullClass_AssertionError() throws Exception {
-        thrown.expect(AssertionError.class);
-        XmlUtil.saveDataToFile(VALID_FILE, null);
-    }
-
-    @Test
-    public void saveDataToFile_missingFile_FileNotFoundException() throws Exception {
-        thrown.expect(FileNotFoundException.class);
-        XmlUtil.saveDataToFile(MISSING_FILE, new AddressBook());
-    }
-
-    @Test
-    public void saveDataToFile_validFile_dataSaved() throws Exception {
-        TEMP_FILE.createNewFile();
-        StorageAddressBook dataToWrite = new StorageAddressBook(new AddressBook());
-        XmlUtil.saveDataToFile(TEMP_FILE, dataToWrite);
-        StorageAddressBook dataFromFile = XmlUtil.getDataFromFile(TEMP_FILE, StorageAddressBook.class);
-        assertEquals((new AddressBook(dataToWrite)).toString(),(new AddressBook(dataFromFile)).toString());
-        //TODO: use equality instead of string comparisons
-
-        AddressBookBuilder builder = new AddressBookBuilder(new AddressBook());
-        dataToWrite = new StorageAddressBook(builder.withPerson("John", "Doe").withTag("Friends").build());
-
-        XmlUtil.saveDataToFile(TEMP_FILE, dataToWrite);
-        dataFromFile = XmlUtil.getDataFromFile(TEMP_FILE, StorageAddressBook.class);
-        assertEquals((new AddressBook(dataToWrite)).toString(),(new AddressBook(dataFromFile)).toString());
-    }
-}
